requirements.txt.txt
streamlit
mysql-connector-python
pandas
matplotlib
seaborn
nltk
wordcloud
textblob
